Split out abstract tests
